+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Added
 - create file from editor
+- more empty files in different languages
 
 ## 6.3.1
 ## Added