+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 - create file from editor
 - more empty files in different languages
 
+## Changed
+- open a version from the history of supported formats only
+
 ## 6.3.1
 ## Added
 - viewing a locked file